+++ nens_auth_client/tests/test_backends.py
@@ -23,7 +23,7 @@ def test_remoteuser_exists(user_getter):
     user_getter.return_value = User(username="testuser")
 
     user = backends.RemoteUserBackend().authenticate(
-        request=None, userinfo={"sub": "remote-uid"}
+        request=None, claims={"sub": "remote-uid"}
     )
     assert user.username == "testuser"
     user_getter.assert_called_with(remote__external_user_id="remote-uid")
@@ -33,7 +33,7 @@ def test_remoteuser_not_exists(user_getter):
     user_getter.side_effect = ObjectDoesNotExist
 
     user = backends.RemoteUserBackend().authenticate(
-        request=None, userinfo={"sub": "remote-uid"}
+        request=None, claims={"sub": "remote-uid"}
     )
     assert user is None
     user_getter.assert_called_with(remote__external_user_id="remote-uid")
@@ -44,7 +44,7 @@ def test_emailverified_exists(user_getter):
 
     user = backends.EmailVerifiedBackend().authenticate(
         request=None,
-        userinfo={
+        claims={
             "sub": "remote-uid",
             "email": "a@b.com",
             "email_verified": True,
@@ -59,7 +59,7 @@ def test_emailverified_not_exists(user_getter):
 
     user = backends.EmailVerifiedBackend().authenticate(
         request=None,
-        userinfo={
+        claims={
             "sub": "remote-uid",
             "email": "a@b.com",
             "email_verified": True,
@@ -74,7 +74,7 @@ def test_emailverified_multiple_exist(user_getter):
 
     user = backends.EmailVerifiedBackend().authenticate(
         request=None,
-        userinfo={
+        claims={
             "sub": "remote-uid",
             "email": "a@b.com",
             "email_verified": True,
@@ -86,7 +86,7 @@ def test_emailverified_multiple_exist(user_getter):
 
 
 @pytest.mark.parametrize(
-    "userinfo",
+    "claims",
     [
         {"sub": "remote-uid", "email": "a@b.com", "email_verified": False},
         {"sub": "remote-uid", "email": "a@b.com"},
@@ -94,9 +94,9 @@ def test_emailverified_multiple_exist(user_getter):
         {"sub": "remote-uid", "email_verified": True},
     ],
 )
-def test_emailverified_no_verified_email(user_getter, userinfo):
+def test_emailverified_no_verified_email(user_getter, claims):
     user = backends.EmailVerifiedBackend().authenticate(
-        request=None, userinfo=userinfo
+        request=None, claims=claims
     )
     assert user is None
     assert not user_getter.called